refactor(ci): share Release Please provenance lookup
ci-gate now checks the provenance run repository against the base repo (fail-closed).
